+++ src/index.ts
@@ -1074,7 +1074,7 @@ export default class Elysia<
 							  })
 				>
 			>
-		) => MaybePromise<Resolver>
+		) => MaybePromise<Resolver | void>
 	): Type extends 'global'
 		? Elysia<
 				BasePath,
@@ -1142,7 +1142,7 @@ export default class Elysia<
 	 *         }
 	 *     }))
 	 */
-	resolve<const Resolver extends Record<string, unknown>>(
+	resolve<const Resolver extends Record<string, unknown> | void>(
 		resolver: (
 			context: Prettify<
 				Context<
@@ -1156,7 +1156,7 @@ export default class Elysia<
 					BasePath
 				>
 			>
-		) => MaybePromise<Resolver>
+		) => MaybePromise<Resolver | void>
 	): Elysia<
 		BasePath,
 		Scoped,
@@ -1201,7 +1201,7 @@ export default class Elysia<
 				},
 				BasePath
 			>
-		) => MaybePromise<NewResolver>
+		) => MaybePromise<NewResolver | void>
 	): Elysia<
 		BasePath,
 		Scoped,
@@ -1248,7 +1248,7 @@ export default class Elysia<
 									Volatile['resolve']
 						  })
 			>
-		) => MaybePromise<NewResolver>
+		) => MaybePromise<NewResolver | void>
 	): Type extends 'global'
 		? Elysia<
 				BasePath,
@@ -4919,7 +4919,7 @@ export default class Elysia<
 	 *         }
 	 *     }))
 	 */
-	derive<const Derivative extends Record<string, unknown>>(
+	derive<const Derivative extends Record<string, unknown> | void>(
 		transform: (
 			context: Prettify<
 				Context<
@@ -4968,7 +4968,7 @@ export default class Elysia<
 	 *     }))
 	 */
 	derive<
-		const Derivative extends Record<string, unknown>,
+		const Derivative extends Record<string, unknown> | void,
 		const Type extends LifeCycleType
 	>(
 		options: { as?: Type },